Fix to support mainline kexec dtb
Mainline kernel passes device tree blob in r2, not in ATAGs. Also the
kexec_mach_type should be 0xffffffff for the mainline kernel for device
tree based booting.

We want to use mainline kernel as is with dtb file and initramfs,
so let's do minimal fixes to support passing dtb in r2. And let's
pass the generic DT machine as the machine type. This allows us to
use the current mainline kexec-tools too.

So basically this patch is just a backport of the current mainline
kernel kexec code that updates the kexec_mach_type based on what kind
of device tree is detected. This means we can kexec boot both mainline
kernels with standard dtb, and legacy kernels using custom Motorola
little-endian dtb.

When attempting to kexec boot legacy kernels, note that probably the
stock kernel won't work for kexec properly. And note that the Lineage
OS initrd seems to depend on SafeStrap config being there for loopback
devices, and needs to use the custom legacy kexec binary with atags.